Begin refactoring Slack code
